refactor: use correct type for shells
