+++ apps/dh/app-dh/src/app/datahub-app.module.ts
@@ -16,7 +16,7 @@
  */
 import { NgModule } from '@angular/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { DhAppCoreFeatureShellModule } from '@energinet-datahub/dh/core/feature-shell';
+import { DhAppCoreFeatureShellModule } from '@energinet-datahub/dh/core/shell-core';
 
 import { DataHubAppComponent } from './datahub-app.component';
