--- libs/dh/core/feature-shell/project.json
@@ -1,26 +1,0 @@
-{
-  "projectType": "library",
-  "root": "libs/dh/core/feature-shell",
-  "sourceRoot": "libs/dh/core/feature-shell/src",
-  "targets": {
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/dh/core/feature-shell"],
-      "options": {
-        "jestConfig": "libs/dh/core/feature-shell/jest.config.js",
-        "passWithNoTests": true
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "options": {
-        "lintFilePatterns": [
-          "libs/dh/core/feature-shell/src/**/*.ts",
-          "libs/dh/core/feature-shell/src/**/*.html"
-        ]
-      },
-      "outputs": ["{options.outputFile}"]
-    }
-  },
-  "tags": ["product:dh", "domain:core", "type:feature"]
-}
