+++ libs/dh/core/shell-core/.eslintrc.json
