+++ libs/dh/core/shell-core/project.json
@@ -1,0 +1,26 @@
+{
+  "projectType": "library",
+  "root": "libs/dh/core/shell-core",
+  "sourceRoot": "libs/dh/core/shell-core/src",
+  "targets": {
+    "test": {
+      "executor": "@nrwl/jest:jest",
+      "outputs": ["coverage/libs/dh/core/shell-core"],
+      "options": {
+        "jestConfig": "libs/dh/core/shell-core/jest.config.js",
+        "passWithNoTests": true
+      }
+    },
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "options": {
+        "lintFilePatterns": [
+          "libs/dh/core/shell-core/src/**/*.ts",
+          "libs/dh/core/shell-core/src/**/*.html"
+        ]
+      },
+      "outputs": ["{options.outputFile}"]
+    }
+  },
+  "tags": ["product:dh", "domain:core", "type:shell"]
+}
