+++ libs/dh/core/shell-core/src/lib/dh-shell-core.module.ts
