+++ libs/dh/core/shell-core/src/lib/shell/shell.component.ts
