+++ libs/dh/core/shell-core/src/lib/shell/shell.module.ts
