+++ libs/ett/auth/data-access-security/src/lib/authentication/ett-authentication.interceptor.spec.ts
@@ -11,7 +11,7 @@ import {
 import { Component } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { ettAuthRoutePath } from '@energinet-datahub/ett/auth/feature-shell';
+import { ettAuthRoutePath } from '@energinet-datahub/ett/auth/shell-auth';
 import { lastValueFrom } from 'rxjs';
 
 import {
