+++ libs/ett/auth/feature-login/src/lib/ett-authentication-link.component.ts
@@ -22,7 +22,7 @@ import {
 } from '@angular/core';
 import { AuthOidcHttp } from '@energinet-datahub/ett/auth/data-access-api';
 import { AbsoluteUrlGenerator } from '@energinet-datahub/ett/core/util-browser';
-import { ettDashboardRoutePath } from '@energinet-datahub/ett/dashboard/feature-shell';
+import { ettDashboardRoutePath } from '@energinet-datahub/ett/dashboard/shell-dashboard';
 import { LetModule } from '@rx-angular/template';
 import { catchError, map, Observable, of } from 'rxjs';
