--- libs/ett/auth/feature-shell/project.json
@@ -1,27 +1,0 @@
-{
-  "projectType": "library",
-  "root": "libs/ett/auth/feature-shell",
-  "sourceRoot": "libs/ett/auth/feature-shell/src",
-  "prefix": "ett",
-  "targets": {
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/ett/auth/feature-shell"],
-      "options": {
-        "jestConfig": "libs/ett/auth/feature-shell/jest.config.js",
-        "passWithNoTests": true
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "options": {
-        "lintFilePatterns": [
-          "libs/ett/auth/feature-shell/src/**/*.ts",
-          "libs/ett/auth/feature-shell/src/**/*.html"
-        ]
-      },
-      "outputs": ["{options.outputFile}"]
-    }
-  },
-  "tags": ["product:ett", "domain:auth", "type:feature"]
-}
