+++ libs/ett/auth/shell-auth/.eslintrc.json
