+++ libs/ett/auth/shell-auth/README.md
