+++ libs/ett/auth/shell-auth/src/lib/ett-auth-route-path.ts
