+++ libs/ett/auth/shell-auth/tsconfig.json
