+++ libs/ett/auth/shell-auth/tsconfig.lib.json
