+++ libs/ett/auth/shell-auth/tsconfig.spec.json
