--- libs/ett/dashboard/feature-shell/project.json
@@ -1,27 +1,0 @@
-{
-  "projectType": "library",
-  "root": "libs/ett/dashboard/feature-shell",
-  "sourceRoot": "libs/ett/dashboard/feature-shell/src",
-  "prefix": "ett-dashboard",
-  "targets": {
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/ett/dashboard/feature-shell"],
-      "options": {
-        "jestConfig": "libs/ett/dashboard/feature-shell/jest.config.js",
-        "passWithNoTests": true
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "options": {
-        "lintFilePatterns": [
-          "libs/ett/dashboard/feature-shell/src/**/*.ts",
-          "libs/ett/dashboard/feature-shell/src/**/*.html"
-        ]
-      },
-      "outputs": ["{options.outputFile}"]
-    }
-  },
-  "tags": ["product:ett", "domain:dashboard", "type:feature"]
-}
