+++ libs/ett/dashboard/shell-dashboard/.eslintrc.json
