+++ libs/ett/dashboard/shell-dashboard/project.json
@@ -1,0 +1,27 @@
+{
+  "projectType": "library",
+  "root": "libs/ett/dashboard/shell-dashboard",
+  "sourceRoot": "libs/ett/dashboard/shell-dashboard/src",
+  "prefix": "ett-dashboard",
+  "targets": {
+    "test": {
+      "executor": "@nrwl/jest:jest",
+      "outputs": ["coverage/libs/ett/dashboard/shell-dashboard"],
+      "options": {
+        "jestConfig": "libs/ett/dashboard/shell-dashboard/jest.config.js",
+        "passWithNoTests": true
+      }
+    },
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "options": {
+        "lintFilePatterns": [
+          "libs/ett/dashboard/shell-dashboard/src/**/*.ts",
+          "libs/ett/dashboard/shell-dashboard/src/**/*.html"
+        ]
+      },
+      "outputs": ["{options.outputFile}"]
+    }
+  },
+  "tags": ["product:ett", "domain:dashboard", "type:shell"]
+}
