+++ libs/ett/dashboard/shell-dashboard/tsconfig.json
