+++ libs/ett/dashboard/shell-dashboard/tsconfig.lib.json
