+++ libs/ett/dashboard/shell-dashboard/tsconfig.spec.json
