+++ workspace.json
@@ -3,18 +3,18 @@
   "projects": {
     "api-dh": "apps/dh/api-dh",
     "app-dh": "apps/dh/app-dh",
-    "app-dh-core-feature-shell": "libs/dh/core/feature-shell",
+    "app-dh-core-feature-shell": "libs/dh/core/shell-core",
     "app-ett": "apps/ett/app-ett",
     "e2e-dh": "apps/dh/e2e-dh",
     "e2e-ett": "apps/ett/e2e-ett",
     "ett-auth-data-access-api": "libs/ett/auth/data-access-api",
     "ett-auth-data-access-security": "libs/ett/auth/data-access-security",
     "ett-auth-feature-login": "libs/ett/auth/feature-login",
-    "ett-auth-feature-shell": "libs/ett/auth/feature-shell",
+    "ett-auth-feature-shell": "libs/ett/auth/shell-auth",
     "ett-core-environments": "libs/ett/core/environments",
     "ett-core-shell": "libs/ett/core/shell-core",
     "ett-core-util-browser": "libs/ett/core/util-browser",
-    "ett-dashboard-feature-shell": "libs/ett/dashboard/feature-shell",
+    "ett-dashboard-feature-shell": "libs/ett/dashboard/shell-dashboard",
     "gf-test-util-matchers": "libs/gf/test-util-matchers",
     "gf-test-util-staging": "libs/gf/test-util-staging",
     "gf-util-browser": "libs/gf/util-browser",
